TS_AI_Django_Starter/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.4.0
